+++ tests/testthat/test-fetchResults.R
@@ -8,6 +8,7 @@ if (DatabaseConnector:::is_installed("ParallelLogger")) {
 
 for (testServer in testServers) {
   test_that(addDbmsToLabel("Fetch results", testServer), {
+   
     connection <- connect(testServer$connectionDetails)
     on.exit(disconnect(connection))
     sql <- "SELECT COUNT(*) AS row_count FROM @cdm_database_schema.vocabulary"
@@ -39,6 +40,17 @@ for (testServer in testServers) {
     )
     expect_gt(dplyr::collect(andromeda$test2)$rowCount[1], 1)
     Andromeda::close(andromeda)
+    
+    skip_if_not(testServer$connectionDetails$dbms %in% c("sqlite", "duckdb"))
+    # dbFetch only fetches n rows
+    sql <- "SELECT * FROM @cdm_database_schema.vocabulary;"
+    sql <- SqlRender::render(sql,
+                             cdm_database_schema = testServer$cdmDatabaseSchema)
+    resultSet <- dbSendQuery(connection, sql)
+    data <- dbFetch(resultSet, n = 1)
+    expect_false(dbHasCompleted(resultSet))
+    expect_equal(nrow(data), 1)
+    dbClearResult(resultSet)
   })
 }