+++ NEWS.md
@@ -1,3 +1,13 @@
+DatabaseConnector 6.3.1
+=======================
+
+Bugfixes:
+
+1. Fixed `dbFetch()` for DBI drivers, no longer ignoring `n` argument.
+
+2. Fix bulk import for Postgres on MacOs.
+
+
 DatabaseConnector 6.3.0
 =======================